+++ src/components/pages/diary/diaryflow/Content/index.ts
@@ -1,0 +1,5 @@
+export { default as Write } from "./Write";
+export { default as Style } from "./Style";
+export { default as Review } from "./Review";
+export { default as Modify } from "./Modify";
+export { default as Draw } from "./Draw";
